feat(stage-2): целостность данных и финансов из аудита
Деньги Float -> Numeric(12,2)/Decimal:
- 11 денежных колонок в 5 таблицах (reservations, rentals, payments, users.balance, balance_history) + миграция b3e1c7d9a240 (downgrade — double precision, не real)
- financial_service: to_decimal/TWO_PLACES/quantize HALF_UP; PriceDetails Decimal; расчёты цены/ставки/штрафа/кредита/остатка на Decimal; конвертация float() на границе Pydantic-схем (контракт API не изменился)
- balance_service, rental_command/rental_update/rental_return, rental_query_service, balance_history_repository — Decimal-арифметика

Гонки и блокировки:
- increment_usage_counter промокода: условный инкремент (times_used < max_uses) с rowcount — параллельные заказы не проходят лимит
- record_promo_code_usage: конфликт PK под savepoint -> PromoCodeUserUsageLimitError (раньше IntegrityError -> 500 + rollback всей транзакции)
- order_validator: pg_advisory_xact_lock(92001, user_id) перед проверкой лимита резервов (сериализует параллельные создания одного юзера)
- Оптимистичная блокировка резервов: колонка version (миграция c9f2d8e1b451) + условный UPDATE в user/admin update-флоу -> 409 при параллельном изменении
- Промокод при конвертации/продлении больше не отбрасывается молча: доменные ошибки пробрасываются, неожиданные -> 409

Транзакции и пагинация:
- commit -> flush в security_audit_service (логины!), holiday_repository, base_repository.save, settings_service, update_*_end_date — единую транзакцию держит middleware
- count-запросы пагинации без eager-опций; accessory_links -> selectinload (нет декартова произведения и завышенного total); новый тест total=1 при 3 аксессуарах
- update_*_end_date: warning + TODO (потребитель — holiday auto-extend)

Тесты: переписаны 'drops promo' -> 'rejected', новые тесты version-409/promo-гонок (6), обновлены моки bool-контракта
Бэкенд: 1421 passed

+++ RentalApp_FASTAPI/api/models/balance_history.py
@@ -1,6 +1,6 @@
 # api/models/balance_history.py
 
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, Text
+from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, Text, Numeric
 from sqlalchemy.orm import relationship
 from api.database_models import Base
 from datetime import datetime, timezone
@@ -13,7 +13,7 @@ class BalanceHistory(Base):
     # +++ ИЗМЕНЕНИЕ: Изменяем поведение внешнего ключа для сохранения истории транзакций +++
     rental_id = Column(Integer, ForeignKey("rentals.id", ondelete="SET NULL"), nullable=True)
 
-    amount = Column(Float, nullable=False, comment="Сумма операции. Отрицательная для списания, положительная для начисления.")
+    amount = Column(Numeric(12, 2), nullable=False, comment="Сумма операции. Отрицательная для списания, положительная для начисления.")
     operation_type = Column(String, nullable=False, comment="Тип операции (e.g., rental_debit, early_return_credit)")
     description = Column(Text, nullable=True, comment="Описание операции")
